% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_annot_credentials}
\alias{check_annot_credentials}
\title{Checks for correct dataset annotation credentials.}
\usage{
check_annot_credentials()
}
\description{
Checks for correct dataset annotation credentials.
}
