% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_s3meta_credentials}
\alias{check_s3meta_credentials}
\title{Checks for correct AWS S3 metadata credentials.}
\usage{
check_s3meta_credentials()
}
\description{
Checks for correct AWS S3 metadata credentials.
}
