% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_data_provenance.R
\name{commit_data_provenance_batch}
\alias{commit_data_provenance_batch}
\title{Commits a batch of data provenance to REDCap about files in S3}
\usage{
commit_data_provenance_batch(df = NULL)
}
\arguments{
\item{df}{Data frame of uuids and data provenance to commit for a batch of files}
}
\description{
Commits a batch of data provenance to REDCap about files in S3
}
\details{
Note:  NAs or blanks in the data frame will not overwrite existing data in this version.
}
