% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_file_entries.R
\name{delete_file_entries}
\alias{delete_file_entries}
\title{Permanently deletes data provenance in REDCap for files that have been deleted in S3}
\usage{
delete_file_entries(uuid = NULL)
}
\arguments{
\item{uuid}{(Required) UUID of object to delete data provenance for}
}
\description{
Permanently deletes data provenance in REDCap for files that have been deleted in S3
}
