% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deleted_files.R
\name{get_all_deleted_files}
\alias{get_all_deleted_files}
\title{Find all files that used to be in S3 that are marked for deletion of their data provenance}
\usage{
get_all_deleted_files()
}
\value{
A data frame containing metadata about all files marked for deletion of data provenance. Requesting data from a specific prefix can speed up the request.
}
\description{
Find all files that used to be in S3 that are marked for deletion of their data provenance
}
\details{
Requires admin credentials.
}
