% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_annotations.R
\name{get_dataset_annotations}
\alias{get_dataset_annotations}
\title{Queries for dataset annotations}
\usage{
get_dataset_annotations(
  DAG = NULL,
  harmonizedOnly = FALSE,
  evenEmptyCols = FALSE
)
}
\arguments{
\item{DAG}{(Optional) A character vector containing the name(s) of the TGR data access group(s) for which to request data if you belong to multiple.}

\item{harmonizedOnly}{(Optional) Whether you want only the annotations that are harmonized across all projects (TRUE) or a complete data set (FALSE, default).}

\item{evenEmptyCols}{(Optional) Whether you want even the empty fields in your final data frame (TRUE) or if you just want columns where there is at least one value in the resulting dataset (FALSE, default).}
}
\value{
Returns a data frame of annotations.
}
\description{
Queries for dataset annotations
}
\author{
Amy Paguirigan
}
