% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datause_metadata.R
\name{get_datause_metadata}
\alias{get_datause_metadata}
\title{Find data use metadata about files in S3}
\usage{
get_datause_metadata(
  bucket,
  prefix = NULL,
  DAG = NULL,
  allowEmpty = F,
  file_type = NULL
)
}
\arguments{
\item{bucket}{(Required) Bucket to query}

\item{prefix}{(Optional) Bucket prefix to query}

\item{DAG}{(Optional) Data access group(s) to query if you are part of multiple DAG's}

\item{allowEmpty}{(Optional) Returns even empty columns if set to TRUE - use if you are interested in applying new annotations}

\item{file_type}{(Optional) Specific file type(s) to query for, such as "bamCramSam", "variants" or "tabularMatrix"}
}
\value{
A data frame containing data use information about files in S3. Requesting data from a specific prefix can speed up the request.
}
\description{
Find data use metadata about files in S3
}
