% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deleted_files.R
\name{get_deleted_files}
\alias{get_deleted_files}
\title{Find files that used to be in S3 that are marked for deletion of their data provenance}
\usage{
get_deleted_files(bucket = NULL, prefix = NULL, DAG = NULL, file_type = NULL)
}
\arguments{
\item{bucket}{(Required) Bucket to query}

\item{prefix}{(Required) Bucket prefix to query}

\item{DAG}{(Optional) Data access group(s) to query if you are part of multiple DAG's}

\item{file_type}{(Optional) Specific file type(s) to query for, such as "bamCramSam", "variants" or "tabularMatrix"}
}
\value{
A data frame containing metadata about files marked for deletion of data provenance. Requesting data from a specific prefix can speed up the request.
}
\description{
Find files that used to be in S3 that are marked for deletion of their data provenance
}
