% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_file_for_deletion.R
\name{mark_file_for_deletion}
\alias{mark_file_for_deletion}
\title{Marks files in REDCap that have been deleted from S3}
\usage{
mark_file_for_deletion(uuid = NULL)
}
\arguments{
\item{uuid}{(Required) UUID of object to mark for deletion of data provenance}
}
\description{
Marks files in REDCap that have been deleted from S3
}
