% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_credentials}
\alias{set_credentials}
\title{Sets credentials from a file.}
\usage{
set_credentials(path)
}
\arguments{
\item{path}{The path to the file to source that contains appropriately defined environment variables (see requiredCredentials.R for an example).}
}
\description{
Sets credentials from a file.
}
