% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_and_commit_data_provenance.R
\name{tag_and_commit_data_provenance}
\alias{tag_and_commit_data_provenance}
\title{Commits data provenance to REDCap about files in S3 with a unique id}
\usage{
tag_and_commit_data_provenance(
  uuid = NULL,
  bucket_name = NULL,
  object_key = NULL,
  DAG = NULL,
  data_provenance = NULL
)
}
\arguments{
\item{uuid}{(Required) Unique ID of the file in S3}

\item{bucket_name}{(Required) Bucket where the file is stored}

\item{object_key}{(Required) Key of the file in S3 (such as tg/projectA/file.csv)}

\item{DAG}{(Required) Data access group to assign this file to}

\item{data_provenance}{(Required) A named list of character objects that contains the data provenance to commit to REDCap about this file.}
}
\description{
Commits data provenance to REDCap about files in S3 with a unique id
}
