% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_file.R
\name{tag_file}
\alias{tag_file}
\title{Tags new files in S3 with a unique id}
\usage{
tag_file(bucket_name = NULL, object_prefix = NULL)
}
\arguments{
\item{bucket_name}{(Required) Bucket where the file is stored}

\item{object_prefix}{(Required) Prefix of the file in S3 (such as tg/projectA/file.csv)}
}
\value{
Returns uuid of the object tagged.
}
\description{
Tags new files in S3 with a unique id
}
