% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository_annotation_functions.R
\name{tgr_definitions}
\alias{tgr_definitions}
\title{Pull Current TGR Annotation and Value definitions}
\usage{
tgr_definitions()
}
\value{
A data frame containing the current TGR Annotations, Values and their Definitions
}
\description{
Pulls current data about annotations from GitHub for use in annotating molecular data sets in the Repository.
}
\author{
Amy Paguirigan
}
