% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_dataset_id.R
\name{validate_dataset_id}
\alias{validate_dataset_id}
\title{Checks to ensure a dataset_id is not already used and is valid}
\usage{
validate_dataset_id(dataset_id = NULL)
}
\arguments{
\item{dataset_id}{(Required) dataset_id(s) you'd like to check}
}
\value{
Returns NA if all dataset_ids are valid, and an array of invalid dataset_ids if any are invalid
}
\description{
Checks to ensure a dataset_id is not already used and is valid
}
\details{
Unfortunately this currently only works for an admin user.
}
