% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_uuid.R
\name{validate_uuid}
\alias{validate_uuid}
\title{Checks to ensure a uuid is not already used}
\usage{
validate_uuid(uuid = NULL)
}
\arguments{
\item{uuid}{(Required) uuid you'd like to check}
}
\value{
Returns NA if all uuids are valid, and an array of invalid uuids if any are invalid
}
\description{
Checks to ensure a uuid is not already used
}
\details{
Note: Only works for admin users.
}
\keyword{internal}
