% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_provenance.R
\name{get_data_provenance}
\alias{get_data_provenance}
\title{Find user supplied data provenance information about files in S3}
\usage{
get_data_provenance(
  bucket = NULL,
  prefix = NULL,
  uuid = NULL,
  DAG = NULL,
  allowEmpty = F,
  file_type = NULL
)
}
\arguments{
\item{bucket}{Bucket(s) to query}

\item{prefix}{Bucket prefix(es) to query}

\item{uuid}{uuid(s) to get data provenance for}

\item{DAG}{(Optional) Data access group to query if you are part of multiple DAGs}

\item{allowEmpty}{(Optional) Returns even empty columns if set to TRUE - use if you are interested in applying new annotations}

\item{file_type}{(Optional) Specific file type(s) to query for, such as "bamCramSam", "variants" or "tabularMatrix"}
}
\value{
A data frame containing metadata about files in S3. Requesting data from a specific prefix or other options can speed up the request.
}
\description{
Find user supplied data provenance information about files in S3
}
\details{
At least one bucket and/or prefix is required.
}
